Rename `utils` `stringUtils`
These util methods are all related to strings, so let's be explicit
about what it is and avoid generic names for things that aren't indices
