+++ server/utils/stringUtils.ts
